Refactor: 이메일 유효성 검사 함수 모듈화
